+++ source/standalone_examples/api/omni.isaac.debug_draw/rtx_lidar.py
@@ -1,10 +1,11 @@
-# Copyright (c) 2022, NVIDIA CORPORATION.  All rights reserved.
-#
-# NVIDIA CORPORATION and its licensors retain all intellectual property
-# and proprietary rights in and to this software, related documentation
-# and any modifications thereto.  Any use, reproduction, disclosure or
-# distribution of this software and related documentation without an express
-# license agreement from NVIDIA CORPORATION is strictly prohibited.
+__copyright__ = "Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved."
+__license__ = """
+NVIDIA CORPORATION and its licensors retain all intellectual property
+and proprietary rights in and to this software, related documentation
+and any modifications thereto. Any use, reproduction, disclosure or
+distribution of this software and related documentation without an express
+license agreement from NVIDIA CORPORATION is strictly prohibited.
+"""
 
 import carb
 from omni.isaac.kit import SimulationApp
@@ -56,7 +57,7 @@ _, (_, sensor) = omni.kit.commands.execute(
 _, render_product_path = create_hydra_texture([1, 1], sensor.GetPath().pathString)
 
 # Create the post process graph that publishes the render var
-sensors.get_synthetic_data().activate_node_template("RtxSensorCpu" + "DebugDrawPointCloud", 0, [render_product_path])
+sensors.get_synthetic_data().activate_node_template("RtxLidar" + "DebugDrawPointCloud", 0, [render_product_path])
 
 simulation_app.update()
 simulation_app.update()